\name{eigeninv}
\alias{eigeninv}
\docType{package}
\title{
A package for generating matrices that have a given set of eigenvalues
}
\description{
Solves the ``inverse eigenvalue problem'' which is to generate a real-valued matrix that has the specified real eigenvalue spectrum.  It uses the Soules' matrix.  It can generate infinitely many dense matrices, symmetric or asymmetric, with the given set of eigenvalues.  Algorithm can also generate stochastic and doubly stochastic matrices, which have the property that their elements are probability measures and that the row sums and colum sums (for doubly stochastic) are equal to 1. 
}
\details{
\tabular{ll}{
Package: \tab eigeninv\cr
Type: \tab Package\cr
Version: \tab 2011\cr
Date: \tab 2011-08-23\cr
License: \tab GPL-2 or greater\cr
LazyLoad: \tab yes\cr
}
}
\author{
Ravi Varadhan, Johns Hopkins University   \cr
URL:  http://www.jhsph.edu/agingandhealth/People/Faculty_personal_pages/Varadhan.html  \cr
Maintainer:  Ravi Varadhan <rvaradhan@jhmi.edu>
}
\references{
MQ Chen, L Han, and M Neumann, On single and double Soules matrices, \emph{Linear Algebra Appl}, 416, p.88-110, 2006

MT Chu and GH Golub, Inverse Eigenvalue Problems, Oxford University Press, 2005
}
